On branch development Your branch is up to date with 'origin/development'.
Changes to be committed:
	modified:   src/chem_pathway_analysis/i__user_model/convert_reaction_system_file.py
	modified:   src/chem_pathway_analysis/p__data_management/data_update.py

+++ src/chem_pathway_analysis/i__user_model/convert_reaction_system_file.py
@@ -61,7 +61,7 @@ def extract_compound_and_stoichiometry(term:str):
     match = stoichiometry_pattern.match(term)
     split = stoichiometry_pattern.split(term)
     if match:
-        stoichiometry = int(match.group(1)) if match.group(1) else 1
+        stoichiometry = float(match.group(1)) if match.group(1) else 1
         compound = match.group(2)
         return {"compound": compound, "stoichiometry": stoichiometry}
     else:
